requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
mysql-connector-python>=8.2.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0
pandas>=2.1.0
plotly>=5.17.0
typing-extensions>=4.8.0
asyncio>=3.4.3
